+++ web/www/avail-light.js
@@ -1,10 +1,14 @@
-import init, { run } from './avail_light_web.js';
+import init, { run, post_message } from './avail_light_web.js';
 
 (async () => {
-    try {
-        await init();
-        await run("hex", null);
-    } catch (error) {
-        console.error("Error:", error);
-    }
+  try {
+    await init();
+    await run("hex", null);
+  } catch (error) {
+    console.error("Error:", error);
+  }
 })();
+
+self.onmessage = (e) => {
+  post_message(JSON.stringify(e.data));
+};